--- bao-web-app/src/main/java/bao/webapp/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package bao.webapp.repository;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-import bao.webapp.model.User;
-
-@Repository
-public interface UserRepository extends CrudRepository<User, Long> {
-}
