--- bao-web-app/src/main/java/bao/webapp/service/UserService.java
@@ -1,12 +1,0 @@
-package bao.webapp.service;
-
-import java.util.List;
-
-import bao.webapp.model.User;
-
-public interface UserService {
-
-	public List<User> list();
-
-	public User findById(long id);
-}
