--- bao-web-app/src/main/resources/application.properties
@@ -1,13 +1,0 @@
-#logging.level.org.h2.server: DEBUG
-
-# Database
-spring.datasource.driver-class-name=com.mysql.jdbc.Driver
-spring.datasource.url= jdbc:mysql://localhost:3306/test_db_01
-spring.datasource.username=root
-spring.datasource.password=root
-
-spring.jpa.hibernate.ddl-auto=create-drop
-
-# logging
-logging.pattern.console=%d{yyyy-MM-dd HH:mm:ss} %-5level %logger{36} - %msg%n
-logging.level.org.hibernate.SQL=debug
