+++ database/bao-web-app/src/main/java/bao/webapp/repository/UserRepository.java
@@ -1,0 +1,9 @@
+package bao.webapp.repository;
+
+import bao.webapp.model.User;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends CrudRepository<User, Long> {
+}
