+++ database/bao-web-app/src/main/java/bao/webapp/service/UserService.java
@@ -1,0 +1,12 @@
+package bao.webapp.service;
+
+import bao.webapp.model.User;
+
+import java.util.List;
+
+public interface UserService {
+
+    List<User> list();
+
+    User findById(long id);
+}
